+++ enterprise/poa/tests/systemtests/poa_test.go
@@ -26,10 +26,9 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/tidwall/gjson"
 
-	"cosmossdk.io/systemtests"
-
 	"github.com/cosmos/cosmos-sdk/crypto/keys/ed25519"
 	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
+	"github.com/cosmos/cosmos-sdk/testutil/systemtests"
 	"github.com/cosmos/cosmos-sdk/types"
 )